修改UI结构
