+++ Android/app/build.gradle
@@ -84,7 +84,7 @@ dependencies {
     //没有使用特殊Header，可以不加这行
     compile 'com.scwang.smartrefresh:SmartRefreshHeader:1.0.4-7'
     //RxTools工具集
-    compile 'com.github.vondear:RxTools:v1.7.7.3'
+    compile 'com.github.vondear:RxTools:v1.7.8.3'
     //图片剪切工具
     compile 'com.yalantis:ucrop:2.2.0'
     //动画库
