+++ Android/app/src/main/java/com/culturer/yoo_home/function/main/MainActivity.java
@@ -1,5 +1,6 @@
 package com.culturer.yoo_home.function.main;
 
+import android.Manifest;
 import android.app.AlertDialog;
 import android.content.DialogInterface;
 import android.content.Intent;
@@ -13,6 +14,7 @@ import android.support.v4.view.ViewPager;
 import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.ActionBarDrawerToggle;
 import android.support.v7.app.AppCompatActivity;
+import android.util.Log;
 import android.view.KeyEvent;
 import android.view.LayoutInflater;
 import android.view.MenuItem;
@@ -21,6 +23,7 @@ import android.widget.LinearLayout;
 import android.widget.Toast;
 
 import com.culturer.yoo_home.R;
+import com.culturer.yoo_home.function.chat.ChatActivity;
 import com.culturer.yoo_home.function.home.HomeFragment;
 import com.culturer.yoo_home.function.mine.MineFragment;
 import com.culturer.yoo_home.function.setting.family_manager.FamilyManagerActivity;
@@ -29,12 +32,16 @@ import com.culturer.yoo_home.function.setting.privacy_setting.PrivacySettingActi
 import com.culturer.yoo_home.function.setting.safe_setting.SafeSettingActivity;
 import com.culturer.yoo_home.function.world.WorldFragment;
 import com.culturer.yoo_home.widget.navigation.impl.HomeNavigation;
+import com.samanlan.lib_permisshelper.PermissionsListener;
+import com.samanlan.lib_permisshelper.PermissionsUtils;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class MainActivity extends AppCompatActivity implements NavigationView.OnNavigationItemSelectedListener  {
 
+    private static final String TAG = "MainActivity";
+    
     private View contentView;
     private List<Fragment> fragments;
     BottomNavigationView navigation;
@@ -68,6 +75,7 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
     }
 
     private void init(){
+        initGrant();
         initData();
         initView();
     }
@@ -198,6 +206,32 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
         return true;
         
     }
+    
+    //授权
+    private void initGrant(){
+        PermissionsUtils mPermissionsUtils = new PermissionsUtils();
+        mPermissionsUtils.setPermissionsListener(new PermissionsListener() {
+            @Override
+            public void onDenied(String[] deniedPermissions) {
+                for (String deniedPermission : deniedPermissions) {
+                    Log.i(TAG, "onDenied: "+deniedPermission + " 权限被拒绝");
+                }
+            }
+            @Override
+            public void onGranted() {
+                Log.i(TAG, "onGranted: 所有权限都被同意");
+            }
+        })
+                .withActivity(this)
+                .getPermissions(MainActivity.this
+                        , 100
+                        , Manifest.permission.RECORD_AUDIO
+                        , Manifest.permission.READ_EXTERNAL_STORAGE
+                        , Manifest.permission.READ_CALENDAR
+                        , Manifest.permission.ACCESS_FINE_LOCATION
+                        ,Manifest.permission.CAMERA);
+    }
+    
     
     @Override
     public void onBackPressed() {
