+++ Android/app/src/main/java/com/culturer/yoo_home/function/setting/create_home/CreateHomeActivity.java
@@ -29,5 +29,4 @@ public class CreateHomeActivity extends AppCompatActivity {
 				build();
 	}
 	
-	
 }
