+++ Android/app/src/main/java/com/culturer/yoo_home/function/world/shows/show/ShowActivity.java
@@ -127,11 +127,5 @@ public class ShowActivity extends AppCompatActivity {
         fragment_container.setAdapter(adapter);
         switchTab(0);
     }
-    @Override
-    public void onBackPressed() {
-        if (!BackHandlerHelper.handleBackPress(this)) {
-            super.onBackPressed();
-        }
-    }
     
 }
