+++ Android/app/src/main/res/layout/activity_setting.xml
@@ -252,141 +252,7 @@
                 android:layout_height="@dimen/divider"
                 android:background="@color/grey"
                 android:layout_marginBottom="@dimen/margin"/>
-            <View
-                android:layout_width="match_parent"
-                android:layout_height="10dp"
-                android:background="@color/grey" />
-            <TextView
-                android:text="基本设置"
-                android:textSize="20dp"
-                android:textColor="@color/black"
-                android:gravity="center_horizontal"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content" />
-            <View
-                android:layout_width="match_parent"
-                android:layout_height="10dp"
-                android:background="@color/grey" />
-            <TextView
-                android:id="@+id/setting_manage_home"
-                android:text="管理家庭"
-                android:textColor="@color/black"
-                android:textSize="@dimen/content_large_textSize"
-                android:gravity="center_vertical"
-                android:layout_margin="@dimen/margin"
-                android:padding="@dimen/padding"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:drawableEnd="@drawable/ic_navigate_next_black_24dp"/>
-            <View
-                android:layout_width="match_parent"
-                android:layout_height="@dimen/divider"
-                android:background="@color/grey"/>
-            <TextView
-                android:id="@+id/setting_manage_family"
-                android:text="管理家族"
-                android:textColor="@color/black"
-                android:textSize="@dimen/content_large_textSize"
-                android:gravity="center_vertical"
-                android:layout_margin="@dimen/margin"
-                android:padding="@dimen/padding"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:drawableEnd="@drawable/ic_navigate_next_black_24dp"/>
-            <View
-                android:layout_width="match_parent"
-                android:layout_height="@dimen/divider"
-                android:background="@color/grey"/>
-            <TextView
-                android:id="@+id/setting_manage_mate"
-                android:text="更换伴侣"
-                android:textColor="@color/black"
-                android:textSize="@dimen/content_large_textSize"
-                android:gravity="center_vertical"
-                android:layout_margin="@dimen/margin"
-                android:padding="@dimen/padding"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:drawableEnd="@drawable/ic_navigate_next_black_24dp"/>
-            <View
-                android:layout_width="match_parent"
-                android:layout_height="@dimen/divider"
-                android:background="@color/grey"/>
-            <TextView
-                android:id="@+id/setting_manage_self"
-                android:text="隐私设置"
-                android:textColor="@color/black"
-                android:textSize="@dimen/content_large_textSize"
-                android:gravity="center_vertical"
-                android:layout_margin="@dimen/margin"
-                android:padding="@dimen/padding"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:drawableEnd="@drawable/ic_navigate_next_black_24dp"/>
-            <View
-                android:layout_width="match_parent"
-                android:layout_height="@dimen/divider"
-                android:background="@color/grey"/>
-            <TextView
-                android:id="@+id/setting_manage_safe"
-                android:text="安全设置"
-                android:textColor="@color/black"
-                android:textSize="@dimen/content_large_textSize"
-                android:gravity="center_vertical"
-                android:layout_margin="@dimen/margin"
-                android:padding="@dimen/padding"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:drawableEnd="@drawable/ic_navigate_next_black_24dp"/>
-            <View
-                android:layout_width="match_parent"
-                android:layout_height="@dimen/divider"
-                android:background="@color/grey"/>
-            <TextView
-                android:id="@+id/setting_manage_password"
-                android:text="修改密码"
-                android:textColor="@color/black"
-                android:textSize="@dimen/content_large_textSize"
-                android:gravity="center_vertical"
-                android:layout_margin="@dimen/margin"
-                android:padding="@dimen/padding"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:drawableEnd="@drawable/ic_navigate_next_black_24dp"/>
-            <View
-                android:layout_width="match_parent"
-                android:layout_height="@dimen/divider"
-                android:background="@color/grey"/>
-            <TextView
-                android:id="@+id/setting_manage_tel"
-                android:text="修改电话"
-                android:textColor="@color/black"
-                android:textSize="@dimen/content_large_textSize"
-                android:gravity="center_vertical"
-                android:layout_margin="@dimen/margin"
-                android:padding="@dimen/padding"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:drawableEnd="@drawable/ic_navigate_next_black_24dp"/>
-            <View
-                android:layout_width="match_parent"
-                android:layout_height="@dimen/divider"
-                android:background="@color/grey"/>
-            <TextView
-                android:id="@+id/setting_manage_email"
-                android:text="修改邮箱"
-                android:textColor="@color/black"
-                android:textSize="@dimen/content_large_textSize"
-                android:gravity="center_vertical"
-                android:layout_margin="@dimen/margin"
-                android:padding="@dimen/padding"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:drawableEnd="@drawable/ic_navigate_next_black_24dp"/>
-            <View
-                android:layout_width="match_parent"
-                android:layout_height="@dimen/divider"
-                android:background="@color/grey"/>
+
         </LinearLayout>
     </ScrollView>
 </LinearLayout>
